% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_xlsx_reader.R
\name{my_xlsx_reader}
\alias{my_xlsx_reader}
\title{Title my_xlsx_reader}
\usage{
my_xlsx_reader(folder = getwd())
}
\arguments{
\item{folder}{}
}
\value{
an XLSX file
}
\description{
Title my_xlsx_reader
}
