% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Title save_as_csv}
\usage{
save_as_csv(x, filedirectory)
}
\arguments{
\item{x}{is the data frame}

\item{filedirectory}{is the name of the CSV we want to create}
}
\value{
an CSV file
}
\description{
Title save_as_csv
}
